drop the blaze-html dependency
Mostly I wanted to get rid of the messy rendering module that was
forked from blaze-html. That library isn't particularly interested in
producing nice-looking html files, so it was slightly at odds with what
I want to do here.

But I also think there's some benefit to having the HTML subset now
defined in the HtmlTypes module. Most importantly, we have types that
distinguish inline from block content, which turns out to be a really
important distinction for two reasons:

 - for writing composable html-generating functions in Haskell
 - having some limitations on which elements can make the transition
   between block and inline content helps when designing CSS